split test data for interfaces and abstract classes
