+++ src/LightMock.Generator.Tests/AbstractClass/BasicMethod.abstract.cs
@@ -1,6 +1,6 @@
-﻿namespace LightMock.Generator.Tests.Testee
+﻿namespace LightMock.Generator.Tests.AbstractClass
 {
-    public abstract class AbstractClass
+    public abstract class BasicMethod
     {
         protected abstract void DoSomething(int p);
         public abstract int GetSomething();
