+++ src/LightMock.Generator.Tests/AbstractClass/TooManyInterfacesWarning.class.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace LightMock.Generator.Tests.Interface
+{
+    interface IFirst { }
+
+    abstract class ASomeClass { }
+
+    [GenerateMock]
+    partial class TooManyInterfacesWarning : ASomeClass, IFirst
+    {
+    }
+}
