+++ src/LightMock.Generator.Tests/Interface/EventSource.interface.cs
@@ -1,4 +1,4 @@
-﻿namespace LightMock.Generator.Tests.Testee
+﻿namespace LightMock.Generator.Tests.Interface
 {
     public delegate void EventHandler<T>(object source, T data);
