+++ src/LightMock.Generator.Tests/Interface/EventSourceGenericClass.class.cs
@@ -4,10 +4,10 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace LightMock.Generator.Tests.Testee
+namespace LightMock.Generator.Tests.Interface
 {
     [GenerateMock]
-    sealed partial class NoInterfaceError
+    public partial class EventSourceGenericClass<T> : IEventSourceGenericClass<T>
     {
     }
 }
