+++ src/LightMock.Generator.Tests/Interface/EventSourceMultipleNamespaces.class.cs
@@ -1,6 +1,6 @@
 ﻿using EventNamespace2;
 
-namespace LightMock.Generator.Tests.Testee
+namespace LightMock.Generator.Tests.Interface
 {
     [GenerateMock]
     public partial class EventSourceMultipleNamespaces : IEventSourceMultipleNamespaces
