+++ src/LightMock.Generator.Tests/Interface/EventSourceMultipleNamespaces.interface.cs
