+++ src/LightMock.Generator.Tests/Interface/GenericClassAndGenericInterface.class.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace LightMock.Generator.Tests.Testee
+namespace LightMock.Generator.Tests.Interface
 {
     [GenerateMock]
     public partial class GenericClassAndGenericInterface<T> : IGenericClassAndGenericInterface<T>
