+++ src/LightMock.Generator.Tests/Interface/MultipleNamespaces.class.cs
@@ -1,7 +1,7 @@
 ﻿using System;
-using MultipleNamespaces2;
+using LightMock.Generator.Tests.Interface.MultipleNamespaces2;
 
-namespace LightMock.Generator.Tests
+namespace LightMock.Generator.Tests.Interface
 {
     [GenerateMock]
     public partial class MultipleNamespaces : IMultipleNamespaces
