+++ src/LightMock.Generator.Tests/Interface/TooManyInterfacesWarning.class.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace LightMock.Generator.Tests.Testee
+namespace LightMock.Generator.Tests.Interface
 {
     interface IFirst { }
     interface ISecond { }
